refactor: move API views to app folders
