+++ eproba/apps/users/api/serializers.py
@@ -1,6 +1,5 @@
+from apps.users.models import User
 from rest_framework import serializers
-
-from .models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
