+++ eproba/apps/wiki/api/serializers.py
@@ -1,0 +1,14 @@
+from apps.wiki.models import Folder, Page
+from rest_framework import serializers
+
+
+class FolderSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Folder
+        fields = "__all__"
+
+
+class PageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Page
+        fields = "__all__"
